use chat manager as a class directly
