+++ src/main/java/be/kakumi/kachat/utils/ProtocolLibChatManager.java
@@ -1,12 +1,13 @@
 package be.kakumi.kachat.utils;
 
+import be.kakumi.kachat.models.Channel;
 import be.kakumi.kachat.utils.protocollib.WrapperPlayServerChat;
 import com.comphenix.protocol.wrappers.WrappedChatComponent;
 import org.bukkit.entity.Player;
 
 import java.util.List;
 
-public class ProtocolLibChatManager extends SimpleChatManager {
+public class ProtocolLibChatManager extends ChatManager {
 
     @Override
     public void sendMessage(String message, List<Player> receivers) {
